+++ package.json
@@ -26,10 +26,5 @@
     "cypress": "^13.15.2",
     "got": "^11.8.3",
     "test-listen": "^1.1.0"
-  },
-  "overrides": {
-    "multer": "^1.4.5-lts.1",
-    "express": "^4.21.2",
-    "body-parser": "^1.20.3"
   }
 }
